backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
google-generativeai==0.8.0
sentence-transformers==3.0.0
chromadb>=0.6.0
httpx==0.27.0
python-multipart==0.0.9
Pillow==10.4.0
reportlab==4.2.0
datasets==2.21.0
